Refactored form service
