+++ src/api/VoteMonitor.Api.Form/CommandHandlers/AddFormCommandHandler.cs
@@ -1,26 +1,26 @@
 ﻿using AutoMapper;
 using MediatR;
-
 using System.Threading;
 using System.Threading.Tasks;
+using VoteMonitor.Api.Form.Commands;
 using VoteMonitor.Api.Form.Models;
+using VoteMonitor.Api.Form.Queries;
 using VoteMonitor.Entities;
 
-namespace VoteMonitor.Api.Form.Queries
+namespace VoteMonitor.Api.Form.CommandHandlers
 {
-    public class AddFormQueryHandler :
-        IRequestHandler<AddFormQuery, FormDTO>
+    public class AddFormCommandHandler : IRequestHandler<AddFormCommand, FormDTO>
     {
         private readonly VoteMonitorContext _context;
         private readonly IMapper _mapper;
 
-        public AddFormQueryHandler(VoteMonitorContext context, IMapper mapper)
+        public AddFormCommandHandler(VoteMonitorContext context, IMapper mapper)
         {
             _context = context;
             _mapper = mapper;
         }
 
-        public async Task<FormDTO> Handle(AddFormQuery message, CancellationToken cancellationToken)
+        public async Task<FormDTO> Handle(AddFormCommand message, CancellationToken cancellationToken)
         {
             var newForm = new Entities.Form();
             var formMapper = new FormDbMapper(_context, _mapper);
