+++ src/api/VoteMonitor.Api.Form/Commands/UpdateFormCommand.cs
@@ -1,0 +1,11 @@
+﻿using MediatR;
+using VoteMonitor.Api.Form.Models;
+
+namespace VoteMonitor.Api.Form.Commands
+{
+    public class UpdateFormCommand : IRequest<FormDTO> 
+    {
+        public FormDTO Form { get; set; }
+        public int Id { get; set; }
+    }
+}
