+++ src/api/VoteMonitor.Api.Form/Models/FormDTO.cs
@@ -1,6 +1,4 @@
-﻿using AutoMapper;
-using System.Collections.Generic;
-using VoteMonitor.Entities;
+﻿using System.Collections.Generic;
 
 namespace VoteMonitor.Api.Form.Models
 {
@@ -16,31 +14,4 @@ namespace VoteMonitor.Api.Form.Models
         public int Order { get; set; }
     }
 
-    public class FormProfile : Profile
-    {
-        public FormProfile()
-        {
-            CreateMap<FormDTO, Entities.Form>()
-                .ForMember(dest => dest.FormSections, c => c.MapFrom(src => src.FormSections))
-                .ForMember(dest => dest.Draft, c => c.MapFrom(src => src.Draft));
-
-            CreateMap<FormSectionDTO, FormSection>()
-                .ForMember(dest => dest.Questions,
-                    c => c.MapFrom(src => src.Questions));
-
-            CreateMap<QuestionDTO, Question>()
-                .ForMember(dest => dest.OptionsToQuestions,
-                    c => c.MapFrom(src => src.OptionsToQuestions));
-
-            CreateMap<OptionToQuestionDTO, OptionToQuestion>()
-                .ForMember(dest => dest.Flagged, c => c.MapFrom(src => src.Flagged))
-                .ForMember(dest => dest.Option, c => c.MapFrom(src =>
-                    new Option
-                    {
-                        Text = src.Text,
-                        IsFreeText = src.IsFreeText
-                    }))
-                ;
-        }
-    }
 }
