+++ src/api/VoteMonitor.Api.Form/Models/FormSectionDTO.cs
@@ -1,6 +1,4 @@
-﻿using AutoMapper;
-using System.Collections.Generic;
-using VoteMonitor.Entities;
+﻿using System.Collections.Generic;
 
 namespace VoteMonitor.Api.Form.Models
 {
@@ -17,21 +15,5 @@ namespace VoteMonitor.Api.Form.Models
         public int OrderNumber { get; set; }
 
         public List<QuestionDTO> Questions { get; set; }
-    }
-
-    public class QuestionProfile : Profile
-    {
-        public QuestionProfile()
-        {
-            CreateMap<Question, QuestionDTO>()
-                .ForMember(dest => dest.OptionsToQuestions, c => c.MapFrom(src => src.OptionsToQuestions));
-
-            CreateMap<OptionToQuestion, OptionToQuestionDTO>()
-                .ForMember(dest => dest.Text, c => c.MapFrom(src => src.Option.Text))
-                .ForMember(dest => dest.IsFreeText, c => c.MapFrom(src => src.Option.IsFreeText))
-                .ForMember(dest => dest.OrderNumber, c => c.MapFrom(src => src.Option.OrderNumber))
-                .ForMember(dest => dest.IdOption, c => c.MapFrom(src => src.Id))
-                .ForMember(dest => dest.Flagged, c => c.MapFrom(src => src.Flagged));
-        }
     }
 }
