--- src/api/VoteMonitor.Api.Form/Models/ModelRaspunsDisponibil.cs
@@ -1,10 +1,0 @@
-﻿namespace VoteMonitor.Api.Form.Models
-{
-    public class ModelRaspunsDisponibil
-    {
-        public int IdOptiune { get; set; }
-        public string TextOptiune { get; set; }
-        public bool SeIntroduceText { get; set; }
-
-    }
-}
