--- src/api/VoteMonitor.Api.Form/Models/ModelVersiune.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections.Generic;
-
-namespace VoteMonitor.Api.Form.Models
-{
-    public class ModelVersiune
-    {
-        /// <summary>
-        /// Collection of <see cref="Formulare"/>
-        /// </summary>
-        public List<Entities.Form> Formulare { get; set; }
-    }
-}
