+++ src/api/VoteMonitor.Api.Form/Models/OptionToQuestionDTO.cs
@@ -2,6 +2,7 @@
 {
     public class OptionToQuestionDTO
     {
+        public int Id { get; set; }
         public int IdOption { get; set; }
         public string Text { get; set; }
         public bool IsFreeText { get; set; }
