+++ src/api/VoteMonitor.Api.Form/Models/Options/OptionDTO.cs
@@ -1,0 +1,10 @@
+namespace VoteMonitor.Api.Form.Models.Options
+{
+    public class OptionDTO
+    {
+        public int Id { get; set; }
+        public bool IsFreeText { get; set; }
+        public string Text { get; set; }
+        public string Hint { get; set; }
+    }
+}
