--- src/api/VoteMonitor.Api.Form/Queries/FetchAllOptionsCommand.cs
@@ -1,11 +1,0 @@
-﻿using MediatR;
-using System.Collections.Generic;
-using VoteMonitor.Api.Form.Models;
-
-namespace VoteMonitor.Api.Form.Queries
-{
-    public class FetchAllOptionsCommand : IRequest<List<OptionDto>>
-    {
-
-    }
-}
