+++ src/api/VoteMonitor.Api.Form/Queries/FetchAllOptionsQuery.cs
@@ -1,0 +1,11 @@
+﻿using MediatR;
+using System.Collections.Generic;
+using VoteMonitor.Api.Form.Models.Options;
+
+namespace VoteMonitor.Api.Form.Queries
+{
+    public class FetchAllOptionsQuery : IRequest<List<OptionDTO>>
+    {
+
+    }
+}
