--- src/api/VoteMonitor.Api.Form/Queries/UpdateFormQuery.cs
@@ -1,12 +1,0 @@
-﻿using MediatR;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using VoteMonitor.Api.Form.Models;
-
-namespace VoteMonitor.Api.Form.Queries {
-    public class UpdateFormQuery  : IRequest<FormDTO> {
-        public FormDTO Form { get; set; }
-        public int Id { get; set; }
-    }
-}
