--- src/api/VoteMonitor.Api.Form/Queries/UpdateOptionCommand.cs
@@ -1,15 +1,0 @@
-﻿using MediatR;
-using VoteMonitor.Api.Form.Models;
-
-namespace VoteMonitor.Api.Form.Queries
-{
-    public class UpdateOptionCommand : IRequest<int>
-    {
-        public OptionDto Option { get; }
-
-        public UpdateOptionCommand(OptionDto option)
-        {
-            Option = option;
-        }
-    }
-}
